implemneted reader functionality.

+++ readmore.php
@@ -26,7 +26,7 @@
                             <li> <a href="edit-book.php?id=<?= $_GET['id']; ?>" class="btn btn-primary m-1" role="button">Edit</a></li>
 
                             <!-- delete button -->
-                            <li> <a href="deletebook.php?id=<?= $_GET['id']; ?>" class="btn btn-danger m-1" role="button">Delete</a></li>
+                            <li> <a href="deletebook.php?id=<?= $_GET['id']; ?>" onclick="return confirm('Are you sure you want to delete this item?')" class="btn btn-danger m-1" role="button">Delete</a></li>
                         </div>
                 <?php }
                 }
@@ -77,18 +77,7 @@
         ?>
 </form>
 
-<!-- footer -->
-
-<footer class="bg-dark text-center fixed">
-    <div class="container text-white p-1">
-    <small>&copy; E-Library 2023. Made with &#10084; in  <a href="https://coloredcow.com/"><img style="width:16px" src="ColoredCow-logo-white.png" alt="logo"></a> ColoredCow Tehri. </small>
-    </div>
-</footer>
-
-<!-- bootstrap js -->
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js" integrity="sha384-w76AqPfDkMBDXo30jS1Sgez6pr3x5MlQ1ZAGC+nuZB+EYdgRZgiwxhTBTkF7CXvN" crossorigin="anonymous"></script>
-
-
+<?php include 'partials/footer.php'; ?>
 </body>
 
 </html>
